backend/requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
pytesseract==0.3.10
Pillow==10.1.0
openai==0.28.1
langchain==0.0.340
spacy==3.7.2
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0